Fix outer module wrapping
ghstack-source-id: e9e162e18ab2a748405ccb26bc3de76781b5f804
Pull Request resolved: https://github.com/pytorch/pytorch/pull/94951

Note: This was the issue where BERT_pytorch broke when the skipfiles check was too wide (see comment in code)

The change to allow inlining more stuff from the top level seems to expose us
to more unsupported stuff. Should we add the support, or are we into compiling stuff
we don't wanna compile?  Leaning towards we don't want to compile this since it
came from NNModule.__setattr__ which seems sketchy.

'''
  File "/scratch/whc/work/pytorch/torch/_dynamo/variables/misc.py", line 50, in const_getattr
    type_to_use = self.objvar.python_type()
  File "/scratch/whc/work/pytorch/torch/_dynamo/variables/torch.py", line 171, in python_type
    return super().python_type()
  File "/scratch/whc/work/pytorch/torch/_dynamo/variables/base.py", line 146, in python_type
    raise NotImplementedError(f"{self} has no type")
NotImplementedError: TorchVariable() has no type
'''